\name{cpackagedemo-package}
\alias{cpackagedemo-package}
\alias{cpackagedemo}
\docType{package}
\title{
\packageTitle{cpackagedemo}
}
\description{
\packageDescription{cpackagedemo}
}
\author{
\packageAuthor{cpackagedemo}

Maintainer: \packageMaintainer{cpackagedemo}
}
\keyword{ package }
\seealso{
\code{\link{my_sum}}
}
\examples{
my_sum(runif(10))
}
